refactor: Adjust copyTextField and gateway ns style
