+++ apps/web/src/components/copy-text/style.less
@@ -13,7 +13,7 @@
         min-height: 35px;
         margin-right: @margin-xxs;
         margin-left: @margin-xxs;
-        background-color: var(--gray-4);
+        border-right: 1px solid var(--gray-4);
     }
 
     .@{mui-prefix}InputBase-formControl.@{mui-prefix}-disabled {
